pyproject.toml
[project]
name = "example-pkg-alt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "FarhadGul", email = "your_actual_github_email@example.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "langdetect>=1.0.9",
    "openai-agents>=0.0.16",
    "python-dotenv>=1.1.0",
]

[project.scripts]
example-pkg = "example_pkg:main"

[tool.hatch.build.targets.wheel]
packages = ["src/example_pkg"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
